pyproject.toml
[tool.poetry]
name = "kanjiconv"
version = "0.1.1"
description = "Kanji Converter to Hiragana, Katakana, Roman alphabet"
authors = ["sea-turt1e <h.yamada.bg@gmail.com>"]
repository = "https://github.com/sea-turt1e/kanjiconv"
license = "Apache-2.0"
readme = "README.md"
keywords = ["kanji", "kana", "converter", "sudachi", "propernoun"]

[tool.poetry.dependencies]
python = "^3.11"
sudachipy = ">=0.6.8"
sudachidict-full = "*"

[tool.poetry.group.dev.dependencies]
pytest-watch = "^4.2.0"
pytest-mock = "^3.14.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
